Transition to objects: OK
Transition to objects is complete, overall program is running as before. Now, the improvements will be continued:

- the calcXDot method must be unified with calcPhi (only one function for system dynamics)
- the tanh for the controls must also be unified, in the form of a method;
- the plotSol method must be improved;
- the debug plots for rest and grad must also be improved (mutiple external options);
- user interface: include methods for more user commands.
